Add files via upload
A simple system for renting homes has two target users: the home owner and the home seeker. The system helps the owner rent his property. After the home seeker applies for the property to rent, the property owner has the option of viewing the applicant's information and accepting or rejecting his request.